% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobNorm.R
\name{RobNorm}
\alias{RobNorm}
\title{RobNorm}
\usage{
RobNorm(X.0, gamma.0 = 0.5, tol = 10^(-4), step = 200)
}
\arguments{
\item{X.0}{The expression matrix in log scale.}

\item{gamma.0}{The density exponent parameter gamma, in practice, taking gamma.0 = 0.5 or 1.}

\item{tol}{The tolerance for interations (default: 10^(-4)).}

\item{step}{The step limit (default: 50).}
}
\value{
Normalized expression data
}
\description{
To robustly normalize expression data.
}
\author{
Meng Wang
\email{mengw1@stanford.edu}
}
