% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim.dat.fn}
\alias{sim.dat.fn}
\title{sim.dat.fn}
\usage{
sim.dat.fn(row.frac, col.frac, mu.up, mu.down, n, m, nu.fix = TRUE)
}
\description{
To simulation an expression matrix (details to see paper RobNorm)
}
\author{
Meng Wang
\email{mengw1@stanford.edu}
}
